+++ src/main/java/com/mftplus/demo/model/service/RefundService.java
@@ -25,7 +25,7 @@ public class RefundService {
     }
 
     @Transactional
-    public void remove(Integer id) {
+    public void remove(Long id) {
         Refund refund = entityManager.find(Refund.class, id);
         entityManager.remove(refund);
     }
@@ -37,7 +37,7 @@ public class RefundService {
     }
 
     @Transactional
-    public Refund findById(Integer id) {
+    public Refund findById(Long id) {
         return entityManager.find(Refund.class, id);
     }
 }
